+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -445,6 +445,7 @@ sub build_isf_public_interface_contract {
             't/1276-isf-enum-member-activation-aggregate-params.t',
             't/1277-isf-enum-member-actor-aggregate-params.t',
             't/1278-isf-enum-member-transaction-aggregate-params.t',
+            't/1279-isf-enum-member-inline-drive-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',