+++ t/1144-isf-public-tested-by-metadata-audit.t
@@ -176,6 +176,7 @@ my $expected_tested_by = [
     't/1276-isf-enum-member-activation-aggregate-params.t',
     't/1277-isf-enum-member-actor-aggregate-params.t',
     't/1278-isf-enum-member-transaction-aggregate-params.t',
+    't/1279-isf-enum-member-inline-drive-values.t',
 ];
 
 subtest 'direct ISF tested_by metadata is exact and valid' => sub {